% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{insstats}
\usage{
sample_mean(x)
}
\arguments{
\item{vector}{}
}
\value{
numeric
}
\description{
calculeaza media unui esantion
}
\examples{

d <- rnorm(100, 15, 0.2)
x <- sample_mean(d)
x

}
