% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_variance.R
\name{sample_variance}
\alias{sample_variance}
\title{insstats}
\usage{
sample_variance(x)
}
\arguments{
\item{vector}{}
}
\value{
numeric
}
\description{
calculeaza varianta unui esantion
}
\examples{

d <- rnorm(100, 15, 0.2)
v <- sample_variance(d)
v

}
